Renamed utils -> lib.
